requirements.txt
streamlit
torch
torchvision
pillow
matplotlib
